+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 2.0.0 (...)
+- **BREAKING** add ability to block users, requires DB update (add 
+  `blocked_users` table)
 - **BREAKING** add API calls for integration with `vpn-manage-portal`, 
   configuration file needs to be updated with new `[ApiAuthentication]` 
   section